functions/main.py
"""
Cloud Function for automatic secret rotation.
Triggered by Cloud Scheduler every 90 days.
Rotates secrets in Google Cloud Secret Manager.
"""

import functions_framework
import json
from google.cloud import secretmanager
from google.cloud import logging as cloud_logging
from datetime import datetime
import base64
import os

# Initialize clients (with fallback for local testing)
try:
    secret_client = secretmanager.SecretManagerServiceClient()
    logging_client = cloud_logging.Client()
    logger = logging_client.logger('secret-rotation-function')
    GCP_AVAILABLE = True
except Exception as e:
    # Local/test mode without GCP credentials
    GCP_AVAILABLE = False
    print(f"⚠️  GCP credentials not available: {str(e)}")
    print("Proceeding in test mode\n")

PROJECT_ID = "aiprod-484120"
ROTATION_THRESHOLD_DAYS = 90

SECRETS_TO_ROTATE = [
    "suno-api-key",
    "freesound-api-key",
    "elevenlabs-api-key",
    "firebase-admin-key",
    "sendgrid-api-key"
]


def perform_rotation():
    """
    Core logic for secret rotation.
    Returns dict with rotation status.
    
    Returns:
        dict: Status of rotation operation
    """
    results = {
        "timestamp": datetime.utcnow().isoformat(),
        "secrets_checked": 0,
        "secrets_rotated": 0,
        "secrets_error": 0,
        "details": []
    }
    
    # Check each secret
    for secret_name in SECRETS_TO_ROTATE:
        results["secrets_checked"] += 1
        
        try:
            if not GCP_AVAILABLE:
                # Test mode: simulate secret data
                detail = {
                    "secret": secret_name,
                    "age_days": 45.0,  # Simulated: not yet due for rotation
                    "needs_rotation": False,
                    "action": "MONITOR"
                }
                results["details"].append(detail)
                continue
            
            secret = secret_client.get_secret(
                request={"name": f"projects/{PROJECT_ID}/secrets/{secret_name}"}
            )
            
            # Calculate age
            created_time = secret.created
            age_seconds = (datetime.utcnow().timestamp() - created_time.timestamp())
            age_days = age_seconds / (24 * 3600)
            
            detail = {
                "secret": secret_name,
                "age_days": round(age_days, 1),
                "needs_rotation": age_days > ROTATION_THRESHOLD_DAYS,
                "action": "NONE"
            }
            
            if age_days > ROTATION_THRESHOLD_DAYS:
                # This is where you would implement actual rotation
                # For now, we just log and alert
                detail["action"] = "REQUIRES_ROTATION"
                
                # In production, you would:
                # 1. Generate new secret value
                # 2. Create new version in Secret Manager
                # 3. Update dependent services
                # 4. Verify old version can still be decrypted
                # 5. Mark old version as destroyed after grace period
                
                if GCP_AVAILABLE:
                    logger.log_text(
                        f"Secret {secret_name} is {age_days:.1f} days old and requires rotation",
                        severity='WARNING'
                    )
                
                results["secrets_rotated"] += 1
            else:
                detail["action"] = "MONITOR"
                if GCP_AVAILABLE:
                    logger.log_text(
                        f"Secret {secret_name} is {age_days:.1f} days old - OK",
                        severity='INFO'
                    )
            
            results["details"].append(detail)
            
        except Exception as e:
            results["secrets_error"] += 1
            error_detail = {
                "secret": secret_name,
                "error": str(e),
                "action": "FAILED"
            }
            results["details"].append(error_detail)
            if GCP_AVAILABLE:
                logger.log_text(
                    f"Error checking secret {secret_name}: {str(e)}",
                    severity='ERROR'
                )
    
    # Log summary
    summary = (
        f"Rotation check complete: {results['secrets_rotated']} rotated, "
        f"{results['secrets_checked'] - results['secrets_rotated']} OK, "
        f"{results['secrets_error']} errors"
    )
    if GCP_AVAILABLE:
        logger.log_text(summary, severity='INFO')
    
    return results


@functions_framework.cloud_event
def rotate_secrets(cloud_event) -> None:
    """
    Main function to rotate secrets.
    Triggered by Cloud Pub/Sub message from Cloud Scheduler.
    
    Args:
        cloud_event: Cloud event containing message data
    """
    try:
        # Parse event message
        pubsub_message = base64.b64decode(cloud_event.data["message"]["data"]).decode()
        logger.log_text(f"Rotation trigger received: {pubsub_message}", severity='INFO')
        
        # Perform rotation and log results
        results = perform_rotation()
        logger.log_text(f"Rotation results: {json.dumps(results)}", severity='INFO')
        
    except Exception as e:
        logger.log_text(f"Fatal error in rotation function: {str(e)}", severity='ERROR')


# Local testing function
if __name__ == "__main__":
    # Test the core rotation logic without Cloud Event wrapper
    print("\n=== Testing Secret Rotation Logic ===")
    result = perform_rotation()
    print(json.dumps(result, indent=2))
