functions/requirements.txt
functions-framework==3.5.0
google-cloud-secret-manager==2.16.1
google-cloud-logging==3.8.0
